No longer use straight.el as the default for all packages
Remove the option to use straight.el as he default for all Emacs
packages as GUIX can be used for packages that are only use on
Linux-based systems. This will result in fewer packages being
downloaded to the ~/.emacs.d/straight/ folder when this configuration
is used on other systems such as Windows which cannot be used with
packages such as EXWM or need to compile packages such as pdf-tools.