Use straight only for system indepdendent packages
No longer use straight.el to manage packages for any packages that are
only used on GUIX or Linux systems as these packages can be managed by
the GUIX system and do not need to be cloned for a portable Emacs
setup on Windows.

Remove the Centaur Tabs package as it provides functionality that I do
not use in my workflow.